Added chargeable boolean parameter to ConferenceRoom constructor as conference rooms are chargeable types of room.
Remove boolean chargeable from DiningRoom constructor as this is not a chargeable type of room.

Created manageBedrooms method to Hotel class to manage the list of bedrooms that are available.  It has the capability to add or remove rooms from the list of bedrooms that the Hotel maintains.  This will be used in the check in/check out process to add to or remove the available rooms.

Added checkInGuests method and helper method called findAvailableRoomForGuests.  The check in method uses the availability checker to identify whether there is availability and if so returns the room that can be filled.

Created addGuestToRoom method in the Room superclass.